Add Users index table
First attempt is in SimpleTable; this didn't work. Tutorial
was from [Medium](https://medium.com/@subalerts/create-dynamic-table-from-json-in-react-js-1a4a7b1146ef)

Second attempt is specific to Users, and it finally works!

TIL there is a render that occurs _before_ the component
is mounted, so render logic that depends on ComponentDidMount
may fail for the first render.