Update UserAuthForm to 'know' inputs
Without using React.createRef, the parent UserAuthForm know
has state values pushed from the child elements AuthInputText.
Even better, this is done with a single function, rather
than one for username and one for password. DRY FTW!

The helpful tutorial was from [React](https://reactjs.org/docs/lifting-state-up.html),
and also credit to [StackOverflow](https://stackoverflow.com/a/695053/10067393)
for how to dynamically set the key (username: or password:).

In summary, AuthInputForm passes onInputChange() to its children
as a prop, which updates the state of the **parent** when included
in the child's onChange handler (in addition to the portion which
sets the child's state.value field, so that the input tracks
and displays its own value).

So far, this means that AuthInputForm will log the submitted
username and password (plaintext), which is a good intermiediate
step to pushing this data to the database.