Build PollDisplay using test data
Current version only displays poll header info and title for
each question, where the questions are selected manually
in the component. This is not ideal.

Plan: Restructure the test data in a hierarchy, so that a poll
has a QUESTIONS key containing all questions for that poll,
each question object contains a RESPONSE_OPTIONS key for its
associated responseOptions, etc.

Provided that the data is fetched from the server and returned
in this (sensible?) format, this should work better. Then
components that are parents/children will have data
structured as needed.