Restructure PollTestData ...
TestData now has QUESTIONS as a key for each POLL, so that
the only two top-level keys of DATA are USERS and POLLS.

Then QUESTIONS consists of a collection of question objects,
and each object contains as a key RESPONSE_OPTIONS, which list
the response options associated only with that question.

With this change, show page correctly renders with title,
description, and author for poll.