Restructure response_options as radio and style
ResponseOptions are now housed in a radio-container div,
and the individual response options are radio input items,
collected together using the name associated with the
parent question id.

Styled radio and input elements to look okay for now.

Next: Make the radio inputs play well with the form (which also
doesn't exist yet), and hold values for submission.