Restructure PollDisplay
Each React component in PollDisplay which does not simply
render passed props makes its own fetch call for data.
This is triggered in componentDidMount, which in turn
triggers a dataLoaded true state. This then (hopefully?) triggers
a re-render with the dataLoaded true, which will then render
the next level of children.

In practice, not all fetches occur, and not all 'Loading...'
messages are displayed.

Planning to search for a better practice. At least this is a
good sign that the generic fetch function works.