Update Polls#show api
With this controller, data fetched includes all data
needed to display a poll, its questions, and the response
options, as well as the author. It is organized into four
objects: POLL, AUTHOR, QUESTIONS, and RESPONSE_OPTIONS.
The fourth is an object with keys question_id and
values the array question.response_options.

Need to update PollDisplay.jsx to accept this single-fetch
model.