Update PollDisplay with form and context
PollDisplay is now ready to submit data! Two major changes:

First, he PollDisplay parent element returns a form, rather than
a div. As such, the child elements reference that form
id, and the bottom input radio elements have the necessary
prop to push their state to the top-level PollDisplay
triggered by onChange.

Second, this prop is passed directly to the input component
using React.createContext. As such, the onRadioChange
function is not passed repeatedly from parent to child,
but instead is passed as a value to context.Provider,
and called from context.Consumer.

OnSubmit of this form now logs the state of the poll display,
specifically the response state, which returns a hash of the
form { question_id: response_option_id }, which are (two
of) the necessary fields for creating a Response object.