Add Respondent select to PollsDisplay
This select menu allows the responding user to be selected
from a dropdown as part of the poll submission process.
Currently, this lists **all** users, and has no concern
for logged-in status.

loadData function has also been moved inside the two components
that use it (PollDisplay and RespondentSelector), as it appears
that (under the previous structure with
`loadData = loadData.bind(this)` ) calling loadData inside
the child element RespondentSelector would **override**
this.state.data of the **parent** component.

As RespondentSelector is an immediate child of PollDisplay,
the onSelectChange prop-passing-method can be passed as
a prop directly (i.e., without the use of Context).

Possible bug: setState is async, so the first console.log
of RespondentSelector.state.respondent_id returns null.
In principle, this could mean that a form get submitted
without a respondent_id; in practice the submit click
would not happen quickly enough for this to be a problem.