Update User and Session routes
Routes for Users#create, Users#index, and Session#create now
utilize resources, rather than manual catching and defining
of the routes. This change has been reflected as needed
in UserAuth (for creating new users and logging in an existing
user) as well as in Users (for displaying the index of users).