Build Answers and Responses api
With the goal of creating Response objects from existing polls
the api now supports creation of Responses and Answers (provided
the latter comes with a response_id).

Validations may need to be checked/introduced to make sure
that an Answer object has a __valid__ response_id; i.e.,
Response.find(id: response_id) not null.

Still need to update the PollsDisplay to push the data to the
api endpoints on Submit.

Also need to put in place JS-level validations for required
(currently all) questions before pushing to endpoint.