Implement pushResponse < PollDisplay
Poll responses now persist! Temporary issue was that
the returned value from the first 'build a response' fetch
call didn't need to be translated via data.json();
later the issue was accidentally looking for the
response_id in data.response_id. This didn't work, since data
is a response object, so the proper source for the id is
data.id -> response_id.

The form still retains the fields after submit, so there needs to
be some clear and/or redirect. Perhaps to a response show page
with stats?

Also there is no validation to prevent multiple responses to the
same poll from the same user, and still no real use for
the login page.