Build Response show API
Create API response controller; this will return the
response data, as well as the respondent and answers data.

Created ResponseFetchData which is the result of GET'ing
these API endpoints for temporary data access; will be later
replaced with fetch() call.

Updated config/routes to include answer and response routes,
both at the api level and the app level.

Updated stylesheet so that responses display in the same manner
as polls.

Build PollHeader of ResponseDisplay; copied from PollDisplay, with
the additional goal of transitioning my React work into only
functional components. So far, so good!