Add `fetch` to ResponseDisplay
Addresses #9; also discovered a bug in responses controller,
which went undeteced since the first two answer objects
had the same id as the first two question objects
(1 and 2); we wanted the response catalog to return an answer
hash with keys 'the question that this one answers'
and values 'the response option selected'.

There is a potential extra/early fetch occuring whereby poll
data is fetched before response data is set to state,
so that response_data.RESPONSE.poll_id is undefined, forcing
another fetch. Creating issue for this.