Add api push to Poll#create
Creating a poll and submitting now pushes the poll to the database,
and renders the PollDisplay component.

This is a good (temporary) solution, but recall the Poll#display
renders a version of the form which can take responses.
This is not ideal for viewing current poll creation status.
It is, however, acceptable for now while creating questions
and response options for the poll.

Existing poll display components could be modified to take in a
new (set of) prop(s) which handle whether the poll is 'display-
only' or 'response-ready'.