Add redirect to PollNew
After initial new poll creation (with title, description,
and author selection), the PollNew component redirects
to the (newly-created) QuestionNew component. This component
will render the existing elements of the poll and provide a form
for creating a new question. config/routes, Router, views,
and controller have been added to accommodate this change.