Build CurrentUserContext
With the goal of having a session management system, we need to
be able to pass info about the current user deep to a child
element. To that end, the Router component in App is now
wrapped in a Context.Provider which will house information
about the current user.

Next steps: Build logic to fetch and set the current user.