Protect session_token
First attempt to address #48 ; the front-end fetches all fields
from the users table via Users.all in the controller, but the
component filters out and only keeps the entries corresponding
to the keys of interest (in this case, id, username, and
admin).

This is inefficient; pages that display poll or response
information will also access the session_token
of the author/respondent. Next attempt will be to adjust the
return/render at the controller level.