Add User#filter
Server level method on a user object, called #filter, which
returns the non-sensitive fields (i.e., not session_token
or password_digest), so that the controllers which
pull a user, author, or respondent (or sereral) can
simply call user.filter as needed.

Removed the front-end code which filtered the data __after__
the api call occured.