Improve logout logic
The `CurrentUserContext` passes its initial state as null, so that
it forces a check for current user on first render of `Home`.
Afterwards, the logout action and a fetch returning
no current user will set `currentUser = {}` -- hence the need to check
for both `currentUser && currentUser.username`; an empty object is
truthy.