Fix setCurrentUser in UserAuthDisplay
The issue from c16936cfea5036e364d85d49fcf6a6203432d9af was
not caused by UserAuthDisplay or the way in which the
context variable currentUser was updated and used to set
a prop for AuthButton, but rather in the controller:

The users and session controllers were set to only return id
and session_token fields upon creation, so there was no username
field. The controllers were updated to protect only password_digest.