Added login link to PollDisplay
If a user views a poll but is not logged in, a previous commit
prevented submission of the poll. Now, there is a link to login
at the bottom of the poll in such a situation, and upon
login, the user is redirected to the very same poll!

Unclear whether/how this works with new user registration.