Refactor PollNew
Given current auth pipeline, the AuthorSelector sub-component
of PollNew is no longer needed; either the poll has no author (in
which case login must occur) or the author is the currrent user.

PollField input is now housed in an adjacent file, and includes
a label field for the input. The 'submit' button is no longer
a separate component, as its logic is simple and unlikely to
be reused.

Finally, PollNew now checks for completeness of required field
(title is the only required field which requires manual input), and
prevents submission if it is missing.

Bug to investigate: `currentUser` is not propogating correctly
when a page is navigated to directly. Hotfix pending.