Move CurrentUserContext to Routes
Previously, the first call of CurrentUserContext inside the
Provider (in App) was in the Home component, which is not
a direct parent of other important components. As such, even
if a currentUser __should__ exist, it would not be stored
in the context.

This commit moves this first call (and the fetch to session to
identify the user with the matching session_token) to the
Routes component, which __is__ a parent of all other components.