Add submit validations to PollView
This commit changes the existing Modal component, adds a new
polls controller, and implements front-end validation logic
to PollView.

**Modal**: First,the folder housing the existing Modal component has been
renamed to Modals, reflecting the possibility that this will later
house additional modals. Second, Modal has been refactored to go
beyond its original 'login required' scope. Specifically, it now
takes four inputs (the final input being an object housing default
values for the final two of five input variables) to allow
for more flexibility. The components previously utilizing this modal
(PollNew and QuestionNew) have been updated to reflect this change.

**Polls Controller**: API/V2 namespace has been implemented, currently
only housing the PollsController#show action. It adds to the previous
V1 controller in that it also returns information about the
respondents of a poll, specifically the data field :RESPONDENTS
with keys respondent ids and values the associated response id.
This allows for the changed logic in PollView.

**PollView logic**: Previous commits (86a4667 and 05a17e3)
added logic to detect required questions and adjust their display
if they are left unanswered on submission attempt. This commit
adds to this front-end validation by checking whether the
currentUser has submitted a response. In this case, a modal is
displayed, blocking form input, and linking to the currentUser's
response.