Add ResponseOption#destroy
This commit adds the controller action needed to receive
and respond appropriately to DELETE requests at the
response_option api endpoint. To this end, config/routes
includes the necessary route, and a view file is added.

The controller searches for the response option object. If it
exists and is destroyed, a message to that end is rendered. If
either fail, a single failure message is rendered instead.

This commit makes progress on #64.