Add #destroy to answers controller
This commit adds the ability to destroy an answer object using
an api endpoint. To this end, config/routes is updated to include
the needed :destroy resource for answers, and the answers controller
is updated appropriately.

In addition, this commit updates the #show action of the
answers controller to return an empty object in the case of a
GET answers/:answer_id request for which there is no
corresponding answer. (This behavior is also seen in, at least,
the users controller.)

Finally, this commit makes needed changes to the PollDisplay
component: A past commit (77d5c55) included a front-end validation
which checked whether there was a currentUser, and, if so, whether
that currentUser had submitted a response to the given poll.
Commit 77d5c55 did not actually __set__ the respondentID to
correspond to the id of the currentUser in that case, so that
the response submission pipeline was broken. This commit
fixes this issue.