Add 'delete' column to PollIndex
See #24; this commit adds access to CurrentUserContext
so that the Polls component can check whether there is
a current user and whether that current user is an admin. In this
case, an additional key, `delete` is added.

In the PollRow component, this key corresponds to a row entry
with a button, which will later include a fetch/DELETE call.

This commit also adds missing `propTypes` definitions which are,
at this time, broken; it appears that chaning `arrayOf` calls
does not work as expected.