Update PollsController#build_index_catalog
This commit changes the data returned from the api endpoint
when rendering poll index data. Previously, the json object
returned/rendered by the controller in this case was an array
(produced by the private method `build_index_catalog`)
of arrays, each array being a triple of poll data:
the first entry the poll object, the second the un-protected
author user object (see e198834), and an integer for the number
of responses to the given poll.

This commit modifies `build_index_catalog` as follows:
each poll in the database returns a `poll_data` object which is
appended to the `catalog` array, where `poll_data` is an object
with keys poll, author, and count (author is protected).

It remains to update the components of /PollsIndex to properly
render this new data structure.