Create PollDeleteContext
This commit adds PollDeleteContext, modeled after
CurrentUserContext (see 2fee83a), to hold and pass information
about whether a poll was just deleted. In this case, the
state can be passed up the chain to a parent component
and trigger a re-render of the polls index table.

To best accommodate this addition, the PollRow component file
has been changed to house all functions related
to the PollRow component __inside__ the component.