Fix PollDeleteContext integration
A previous commit (6519f79) added the PollDeleteContext and imported
it into the hightest ancestor of the PollsIndex component tree.
However, this component (Polls) also needed to access the context
information, so it could not also be the source
of the Provider.

This commit restructures the PollsIndex component tree
to place the bulk of the logic in the second-highest
component (PollsTable), which will then sit under the
PollDeleteProvider.

It remains to fix an issue where multiple re-logins of an admin
added multiple 'delete' columns to the table.