Fix 'delete' column `push`
A previous commit (8b4b9ac) moved the bulk of the fetch
and related logic from Polls to PollsTable. Since this included
the origin of the 'keys' prop, and also added 'delete' to the keys,
this caused two problems:
1. the 'delete' column persisted, regardless of admin status, and
2. an additional and identical 'delete' column would be added
each time an admin user logged in.

This commit fixes both issues by returning the setting of the
'keys' array to the parent component Polls, where
currentUser.admin is also checked. Then, on re-login,
the array is reset, appended (as needed), and passed correctly
as a prop to PollsTable.