Update PollsTable rendering
This commit adds an automatic re-render of the PollsTable
in the situation where a poll was just deleted by an admin user.
This is accomplished by reading the context-connected `pollDelete`
boolean, and, in the case that it is true, (e.g., changed
from the default false by the PollRow component when deletion
occurs, NB any change causes a re-render) then a fetch occurs,
even if `data.mounted` is also true.

It remains to show some message confirming the delete
and/or adding a message of the form 'are you sure you want
to delete this poll'.