Update ResponseDisplay for v3 api
This commit updates the ResponseDisplay component to correctly
interact with the new api/v3 endpoint for responses. In addition
the "RESPONSE" key is removed from the compiled catalog, as
its data is either not needed or already stored in a more
helpful form elsewhere in the catalog.

It remains to build logic to interact with the new "RESPONDENTS"
object of the returned data.