Add `GO` to RespondentDisplay
This commit includes logic needed to determine whether a GO button
should be displayed next to the selected respondent. The values
of the options in the select menu correspond response ids, and
text displayed the username of the author of the given response.

As a test, clicking the button (in the case it is displayed) logs
the response id for the displayed author, as intended.

The stylesheet is also updated to display a reasonable button shape, and
the parent component ResponseDisplay passes the response id
as a prop so that RespondentDisplay logic can compare value in the
select menu to the currently viewed response object id.