Change `GO` button to `Link`
The GO button associated with the respondent selector menu
is now a react-router-dom Link component.

This doesn't actually work as intended, as the page renders (which
is good, an a href element didn't do that), but it doesn't
render the correct response. That is, it only renders the same
response that it started as (though it does correctly adjust the URL).