Add `isAdmin` prop to Users
This commit adds the ability to call UsersIndex with (and without)
admin privleges (though in practice, it will not be called at
all unless by admin). Passing in the prop `isAdmin` with value `true`
into Users will render UsersIndex with 4 columns: `id`, `username`,
`admin`, and `delete`.

It remains to add code to display something helpful in this final column
(with a helpful action) and render this table only when
viewing an admin's show page as that admin.