Add length 0 handling to ResponseOptionsContainer
This commit adds a displayed message to the PollDispaly component
in the case that a question has no response options. Note that
the api/v2/polls endpoint returns an empty array for the
response options object associated with the given question id,
so that there will not be an error of attempting to interact
with a null (or undefined) object.

It remains to adjust the required question completeness
front-end validation (see 05a17e3), as this change allows a required
question to be un-answerable.