Remove empty response options display
This commit reverses the length 0 response options handling
in ResponseOptionsContainer (see 4414066) in preparation for
housing functionally similar logic in the parent QuestionDisplay
component. This is possible since the QuestionDisplay already
has the response options as a prop, so there is no need to pass
an empty prop down to ResponseOptionsContainer and perform
the logic there.

Moreover, we can then also determine (without creating an additional
prop) whether the question is required. This will allow for the
addition of logic to detect whether a required question is
un-answerable (due to absence of response options).