Add empty response option logic to QuestionDisplay
This commit includes logic to detect whether a question object
has no response options and, in this case, whether the quesiton
is required. If so, it display a message. Otherwise,
ResponseOptionsContainer is rendered as usual.

It remains to adjust the unanswered required question validation
in the case when a required question has no response options, and
is thus un-answerable.