Update no response options message
This commit updates both the PollDisplay and ResponseDisplay components,
including a Link to the edit poll pipeline (actually the
new poll pipeline; see #44) in the case that a required question
has no response options.

This will not handle well (yet) the case where a non-required
question has no response options.