Add 'no questions' message to ResponseDisplay
This commit makes progress on #40 by adding a message and link
to a response display in the case that the parent poll has no questions,
see b37f9cf. In order to access the id of the parent poll, we
also need access to the response object. As such, this commit also
updates api/v1/responses_controller include the response object
in the returned catalog; this object was removed in 2a13484.