Refactor UserDisplay
This commit is part of #69; we begin by refactoring UserDisplay
and removing components that require minimal logic and/or
do little more than pass props down to the next generation.

To this end, the UserHeader, Activities, and ActivityHeader
components are removed. Functionally, no changes are made
to the UserDisplay component in this commit.