Improve button and UserDisplay styling
The css selector .small-btn now acts as an extension of .submtit-btn,
only changing the padding and margin, allowing the button to
display more naturally in a small space.

The UserDisplay now has the activities container as a flex-row
display property.

Finally, this commit removes un-needed import and input variable
from portions of the RespondentDisplay component.